rename and clean up interfaces.
