+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/MetadataDescribable.java
@@ -1,10 +1,15 @@
 package com.yahoo.bard.webservice.data.config;
 
 /**
- * An object that is exposed as metadata through the metadata endpoints and describable by {@link GlobalMetadata} or one
+ * An object that is exposed as metadata through the metadata endpoints and describable by {@link CommonMetadata} or one
  * of its sub-interfaces.
  */
 public interface MetadataDescribable {
 
-    GlobalMetadata getMetadata();
+    /**
+     * Getter.
+     *
+     * @return the metadata object.
+     */
+    CommonMetadata getMetadata();
 }
